Assignment And Exam Error Handling

+++ app/Http/Controllers/AssignmentController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
 use App\Http\Controllers\Controller as Controller;
+use Illuminate\Support\Facades\Validator;
 use App\Assignment;
 use App\Student;
 
@@ -22,25 +23,35 @@ class AssignmentController extends Controller
     public function create(Request $request, $teacherid){
 
 
-    	$assignmentData = $request->validate([
+    	// $assignmentData = $request->validate([
 
-    			'assignment'=>'required',
-    			'class_id'=>'required'
+    	// 		'assignment'=>'required',
+    	// 		'class_id'=>'required'
 
-    	]);
+    	// ]);
 
-    	$assignment=  new Assignment;
+    	$assignment =  new Assignment;
+        $validator = Validator::make($request->all(),[
+                'assignment'=>'required',
+                'class_id'=>'required'
+        ]);
 
-    	$assignment->description = $request->assignment;
-    	$assignment->teacher_id = $teacherid;
-    	$assignment->class_id = $request->class_id;
+        if ($validator->passes()) 
+        {
+            $assignment->description = $request->assignment;
+            $assignment->teacher_id = $teacherid;
+            $assignment->class_id = $request->class_id;
 
-    	$assignment->save();
+            $assignment->save();
 
-    	$response['message'] = "Assignment Created";
-    	$response['code'] = "201";
+            $response['message'] = "Assignment Created";
+            $response['code'] = "201";
 
-    	return $response;
+            return $response;
+        }
+
+        return response()->json(['error','Please FillOut All The Fields Required']);
+    	
 
     }
 
@@ -55,19 +66,38 @@ class AssignmentController extends Controller
 
     	$assignment = Assignment::where('teacher_id',$teacherid)->paginate(10);
 
-    	$assignment['message'] = "Records Fetched Successfully";
-    	$assignment['code'] = "200";
+        if (count($assignment)) 
+        {
+            $assignment['message'] = "Records Fetched Successfully";
+             $assignment['code'] = "200";
 
-    	return $assignment;
+            return $assignment;
+        }
+
+        return response()->json(['error','No Records Available']);
 
     }
 
     public function displayStudentAssignment($studentid)
     {
         $student = Student::where('id',$studentid)->first();
-        $student_class_id = $student->class_id;
-        $assignment = Assignment::where('class_id',$student_class_id)->paginate();
-        return response()->json(['success'=>$assignment]);
+
+        if (count($student)) 
+        {
+            $student_class_id = $student->class_id;
+            $assignment = Assignment::where('class_id',$student_class_id)->paginate();
+            
+            if (count($assignment)) 
+            {
+                 return response()->json(['success'=>$assignment]);
+            }
+
+           return response()->json(['error','No Assignments Available']);
+        }
+
+        
+         return response()->json(['error','Student Does Not Exist']);
+       
     }
 
     public function update($teacherid,$assignmentid,Request $request){
@@ -80,15 +110,29 @@ class AssignmentController extends Controller
 
     	]);
 
-    	$assignment->description = $request->assignment;
-    	$assignment->teacher_id = $teacherid;
-    	$assignment->class_id = $request->class_id;
+        $validator = Validator::make($request->all(),[
 
-    	$assignment->save();
-    	$assignment['message'] = "Assignment Updated";
-    	$assignment['code'] = "200";
+            'assignment'=>'required',
+            'class_id'=>'required'
 
-    	return $assignment;
+        ]);
+
+        if ($validator->passes()) 
+        {
+            
+            $assignment->description = $request->assignment;
+            $assignment->teacher_id = $teacherid;
+            $assignment->class_id = $request->class_id;
+
+            $assignment->save();
+            $assignment['message'] = "Assignment Updated";
+            $assignment['code'] = "200";
+
+            return $assignment;
+
+        }
+
+    	return response()->json(['error','Assignment Not Found!! ']);
 
 
     }
@@ -97,12 +141,15 @@ class AssignmentController extends Controller
     {
     	$assignment = Assignment::findOrFail($id);
 
-    	$assignment->delete();
+        if (count($assignment)) 
+        {
+            $assignment->delete();
+            $response['message'] = "Assignment Deleted Succesfully";
+            $response['code'] = "200";
+            return $response;
+        }
 
-    	$response['message'] = "Assignment Deleted Succesfully";
-    	// $response['ass-id'] = $id;
-    	$response['code'] = "200";
-
-    	return $response; 
+        return response()->json(['error','Record Not Found']);
+    	 
     }
 }
